% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_four.R
\name{add_four}
\alias{add_four}
\title{Add four to a numeric vector}
\usage{
add_four(x)
}
\arguments{
\item{x}{(numeric) a numeric vector}
}
\value{
a numeric vector
}
\description{
Add four to a numeric vector
}
\examples{
add_four(2)
add_four(4)
}
