% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{Add one to a numeric vector}
\usage{
add_one(x)
}
\arguments{
\item{x}{(numeric) a numeric vector}
}
\value{
a numeric vector
}
\description{
Add one to a numeric vector
}
\examples{
add_one(2)
}
