% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\name{add_two}
\alias{add_two}
\title{Add two to a numeric vector}
\usage{
add_two(x)
}
\arguments{
\item{x}{(numeric) a numeric vector}
}
\value{
a numeric vector
}
\description{
Add two to a numeric vector
}
\examples{
add_two(2)
}
